app/requirements.txt
click==8.0.4
dataclasses==0.8
Flask==2.0.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.30
protobuf==3.19.5
typing_extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
